# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cosmic Webapp",
    "version": "1.0.0",
    "description": "API documentation with instructions"
  },
  "servers": [
    {
      "url": "/cosmic/"
    },
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/matter-model": {
      "get": {
        "summary": "Get all locations",
        "operationId": "getLocations",
        "tags": ["Locations"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Locations"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/synhronize-model": {
      "get": {
        "summary": "Synchronize all locations from Matterport",
        "operationId": "getSynh",
        "tags": ["Locations"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success message or error of synchronization",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Result message of synchronization"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/locations": {
      "post": {
        "summary": "Create a new location",
        "operationId": "createMatterTag",
        "tags": ["Locations"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "requestBody": {
          "description": "Request body for creating a new location.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location_name": {
                    "type": "string",
                    "description": "Name of the location",
                    "example": "Location #1 Name"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrition of location",
                    "example": "Location #1 Description"
                  },
                  "x": {
                    "type": "number",
                    "description": "Position x",
                    "example": "0"
                  },
                  "y": {
                    "type": "number",
                    "description": "Position y",
                    "example": "0"
                  },
                  "z": {
                    "type": "number",
                    "description": "Position z",
                    "example": "0"
                  },
                  "color": {
                    "type": "string",
                    "description": "Color of matter tag",
                    "example": "#000000"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "Is enabled",
                    "example": true
                  },
                  "floorId": {
                    "type": "string",
                    "description": "ID of flour",
                    "example": "tsmq1wak12rhgn0mawksxcwcd"
                  }
                },
                "required": [
                  "location_name",
                  "description",
                  "anchorPosition",
                  "color",
                  "enabled",
                  "floorId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Locations"
                      }
                    },
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all locations with tasks",
        "operationId": "getAllLocations",
        "tags": ["Locations"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Locations"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/locations/{id}": {
      "get": {
        "tags": ["Locations"],
        "summary": "Get location and associated tasks",
        "operationId": "getLocationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the location",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the location and associated tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tag": {
                      "$ref": "#/components/schemas/Locations"
                    },
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a location by tag ID",
        "tags": ["Locations"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the location",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Locations"
                      }
                    },
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ACCESS_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "REFRESH_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer",
            "description": "The unique identifier of the task"
          },
          "title": {
            "type": "string",
            "description": "Title of the task"
          },
          "description": {
            "type": "string",
            "description": "Description of the task"
          },
          "lists_id": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "An array of list IDs this task belongs to. A task can belong to multiple lists."
          },
          "status": {
            "type": "string",
            "description": "Current status of the task"
          },
          "dueDate": {
            "type": "date",
            "description": "Due date task"
          },
          "priority": {
            "type": "integer",
            "description": "Priority of task"
          },
          "activity": {
            "type": "object",
            "description": "Related activity of the task",
            "properties": {
              "activity_id": {
                "type": "integer",
                "description": "ID of activity"
              },
              "name": {
                "type": "string",
                "description": "Name of activity"
              }
            }
          },
          "locations": {
            "type": "object",
            "description": "Location information",
            "properties": {
              "location_id": {
                "type": "string",
                "description": "ID of location"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "description": "Creation timestamp"
              },
              "location_name": {
                "type": "string",
                "description": "Name of location"
              },
              "color": {
                "type": "string",
                "description": "Color of location in HEX"
              },
              "description": {
                "type": "string",
                "description": "Description of location"
              },
              "x": {
                "type": "number",
                "description": "X coordinate"
              },
              "y": {
                "type": "number",
                "description": "Y coordinate"
              },
              "z": {
                "type": "number",
                "description": "Z coordinate"
              },
              "keywords": {
                "type": "string",
                "nullable": true,
                "description": "Location keywords"
              },
              "matterport_tag_id": {
                "type": "string",
                "description": "Matterport Tag ID"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was last updated"
          }
        }
      },
...
      "Locations": {
        "type": "object",
        "properties": {
          "location_name": {
            "type": "string",
            "description": "Name of the location",
            "example": "Location #1 Name"
          },
          "description": {
            "type": "string",
            "description": "Descrition of location",
            "example": "Location #1 Description"
          },
          "x": {
            "type": "number",
            "description": "Position x",
            "example": "0"
          },
          "y": {
            "type": "number",
            "description": "Position y",
            "example": "0"
          },
          "z": {
            "type": "number",
            "description": "Position z",
            "example": "0"
          },
          "color": {
            "type": "string",
            "description": "Color of location",
            "example": "#000000"
          },
          "enabled": {
            "type": "boolean",
            "description": "Is enabled",
            "example": true
          },
          "floorId": {
            "type": "string",
            "description": "ID of flour",
            "example": "tsmq1wak12rhgn0mawksxcwcd"
          }
        }
      },
...
    }
  }
}